画面を含め機能的には概ね実装完

+++ typeorm-sample/src/services/blog-service.ts
@@ -7,6 +7,7 @@ import { Repository, FindOptions } from "typeorm";
 import { OrmRepository } from "typeorm-typedi-extensions";
 import { BadRequestError, NotFoundError } from "routing-controllers";
 import { Blog } from "../entities/blog";
+import { Article } from "../entities/article";
 
 /**
  * ブログサービスクラス。
@@ -16,6 +17,9 @@ export class BlogService {
 	/** ブログエンティティリポジトリ */
 	@OrmRepository(Blog)
 	private blogRepository: Repository<Blog>;
+	/** ブログ記事エンティティリポジトリ */
+	@OrmRepository(Article)
+	private articleRepository: Repository<Article>;
 
 	/**
 	 * ブログ一覧を取得する。
@@ -40,7 +44,7 @@ export class BlogService {
 	async findOneById(id): Promise<Blog> {
 		const blog = await this.blogRepository.findOneById(id);
 		if (!blog) {
-			throw new NotFoundError(`blog id=${id} is not found`);
+			throw new NotFoundError(`blog is not found`);
 		}
 		return blog;
 	}
@@ -53,7 +57,7 @@ export class BlogService {
 	async insert(blog: Blog): Promise<Blog> {
 		let count = await this.blogRepository.count({ title: blog.title });
 		if (count > 0) {
-			throw new BadRequestError(`blog title=${blog.title} is already existed`);
+			throw new BadRequestError(`blog title is already existed`);
 		}
 		return this.blogRepository.persist(blog);
 	}
@@ -66,11 +70,16 @@ export class BlogService {
 	async update(blog: Blog): Promise<Blog> {
 		const old = await this.blogRepository.findOneById(blog.id);
 		if (!old) {
-			throw new NotFoundError(`blog id=${blog.id} is not found`);
+			throw new NotFoundError(`blog is not found`);
 		}
-		let count = await this.blogRepository.count({ title: blog.title });
+		const op: FindOptions = {
+			alias: "blog",
+			where: "title = :title && id != :id",
+			parameters: { title: blog.title, id: blog.id },
+		};
+		let count = await this.blogRepository.count(op);
 		if (count > 0) {
-			throw new BadRequestError(`blog title=${blog.title} is already existed`);
+			throw new BadRequestError(`blog title is already existed`);
 		}
 		return this.blogRepository.persist(Object.assign(old, blog));
 	}
@@ -83,7 +92,11 @@ export class BlogService {
 	async delete(id: number): Promise<Blog> {
 		const blog = await this.blogRepository.findOneById(id);
 		if (!blog) {
-			throw new NotFoundError(`blog id=${id} is not found`);
+			throw new NotFoundError(`blog is not found`);
+		}
+		let count = await this.articleRepository.count({ blogId: id });
+		if (count > 0) {
+			throw new BadRequestError(`blog has articles`);
 		}
 		return this.blogRepository.remove(blog);
 	}
